Update README; update status command
This updates the README to match what's been implemented.

This makes the status command much more useful. You can have it generate
a list of file names which you can send to another command line program
for tweaking. If you pass no arguments, it uses blessings to colorize
the output making it nicer.