Add webedit
webedit allows someone to use their web browser to edit the
json files. It takes advantage of the fact that steve knows about
the structure of the json files and therefore can make it easier on
an editor.

Having said that, it's pretty basic. It isn't all-encompassing and
should be considered one of many tools steve provides to build
conference data.