Issue 154. Fix epically dumb thing
This fixes an epically dumb thing I did. What's going on is that
if the video is in DRAFT, then it only shows up in the API if you're
an admin. So if you try to do a GET on it with no username/api_key,
then richard gets all fussy and says, "Bugger off!" And to make
matters worse, this is the check for making sure that the video
exists before you can update it. Doh!