Flesh out scrapevideo
This makes scrapevideo work better. It does what it did before, but you
can also use it to scrape a single video instead of doing a fetch.

This tweaks push so that it doesn't overwrite by default. It "knows"
that the video has been pushed by looking at the "id" field. That kind
of sucks for bookkeeping, but ... it's "good enough for now".