Prevent webedit from dropping URLs inserting CR
webedit:route_save() failed to handle URLField, thereby dropping all
URLs when saving. This was not detected as unrecognised req types did
not throw an error.

A roundtrip edit would change all newlines to carriage-return + newline
in TextField type reqs. This obfuscates real changes in these fields.
This is an artifact of browsers handling textareas.

It could be that richard cannot handle the newlines only, but that is
unlikely for a Unix based tool.