Remove hardcoded assumption of "json/0123_" names
The util:load_json_files() routine returned a file name as the first
element of the tuples. This name always started with "json/" and
in several places it was assumed:

- that the current directory was the parent directory of "json" (the
  project directory) and that files could be opened by their name
- that the first 5 characters needed stripping when webediting

An extra config entry 'project/jsonpath' is loaded (or created based on
'project/projectpath), and loading/saving is done under this directory.
This means that the json directory could be anywhere as long as
the project configuration correctly specifies where it is.

File names are now passed without any prefix and don't need chomping,
but do always need prefixing by jsonpath when loading/saving.